+++ engine/class_modules/monk/sc_monk.cpp
@@ -36,7 +36,7 @@ BREWMASTER:
 #include "player/pet_spawner.hpp"
 #include "action/action_callback.hpp"
 #include "sc_enums.hpp"
-#include "report/sc_highchart.hpp"
+#include "report/highchart.hpp"
 #include "report/charts.hpp"
 
 #include "simulationcraft.hpp"
