+++ engine/class_modules/sc_mage.cpp
@@ -8,7 +8,7 @@
 #include "util/util.hpp"
 #include "class_modules/apl/mage.hpp"
 #include "report/charts.hpp"
-#include "report/sc_highchart.hpp"
+#include "report/highchart.hpp"
 
 namespace {
