+++ engine/report/report_helper.cpp
@@ -15,7 +15,7 @@
 #include "player/covenant.hpp"
 #include "report/gear_weights.hpp"
 #include "report/charts.hpp"
-#include "report/sc_highchart.hpp"
+#include "report/highchart.hpp"
 #include "sim/sc_sim.hpp"
 #include "sim/scale_factor_control.hpp"
 #include "util/xml.hpp"
