+++ engine/report/report_html_player.cpp
@@ -13,10 +13,10 @@
 #include "report/report_helper.hpp"
 #include "report/decorators.hpp"
 #include "report/charts.hpp"
+#include "report/highchart.hpp"
 #include "player/player_talent_points.hpp"
 #include "player/scaling_metric_data.hpp"
 #include "player/set_bonus.hpp"
-#include "sc_highchart.hpp"
 #include "sim/scale_factor_control.hpp"
 #include "sim/sc_profileset.hpp"
 #include "util/util.hpp"
