+++ engine/report/report_text.cpp
