+++ source_files/QT_engine.pri
@@ -128,11 +128,11 @@ HEADERS += engine/report/charts.hpp
 HEADERS += engine/report/color.hpp
 HEADERS += engine/report/decorators.hpp
 HEADERS += engine/report/gear_weights.hpp
+HEADERS += engine/report/highchart.hpp
 HEADERS += engine/report/json/report_configuration.hpp
 HEADERS += engine/report/report_helper.hpp
 HEADERS += engine/report/report_timer.hpp
 HEADERS += engine/report/reports.hpp
-HEADERS += engine/report/sc_highchart.hpp
 HEADERS += engine/sc_enums.hpp
 HEADERS += engine/sim/benefit.hpp
 HEADERS += engine/sim/cooldown_waste_data.hpp
@@ -302,15 +302,15 @@ SOURCES += engine/player/unique_gear_shadowlands.cpp
 SOURCES += engine/report/charts.cpp
 SOURCES += engine/report/color.cpp
 SOURCES += engine/report/decorators.cpp
+SOURCES += engine/report/gear_weights.cpp
+SOURCES += engine/report/highchart.cpp
 SOURCES += engine/report/json/report_configuration.cpp
 SOURCES += engine/report/json/report_json.cpp
 SOURCES += engine/report/report_helper.cpp
+SOURCES += engine/report/report_html_player.cpp
+SOURCES += engine/report/report_html_sim.cpp
+SOURCES += engine/report/report_text.cpp
 SOURCES += engine/report/reports.cpp
-SOURCES += engine/report/sc_gear_weights.cpp
-SOURCES += engine/report/sc_highchart.cpp
-SOURCES += engine/report/sc_report_html_player.cpp
-SOURCES += engine/report/sc_report_html_sim.cpp
-SOURCES += engine/report/sc_report_text.cpp
 SOURCES += engine/sim/cooldown_waste_data.cpp
 SOURCES += engine/sim/event_manager.cpp
 SOURCES += engine/sim/proc.cpp
