+++ source_files/VS_engine.props
@@ -132,11 +132,11 @@ To change the list of source files run synchronize.py
 		<ClInclude Include="..\engine\report\color.hpp" />
 		<ClInclude Include="..\engine\report\decorators.hpp" />
 		<ClInclude Include="..\engine\report\gear_weights.hpp" />
+		<ClInclude Include="..\engine\report\highchart.hpp" />
 		<ClInclude Include="..\engine\report\json\report_configuration.hpp" />
 		<ClInclude Include="..\engine\report\report_helper.hpp" />
 		<ClInclude Include="..\engine\report\report_timer.hpp" />
 		<ClInclude Include="..\engine\report\reports.hpp" />
-		<ClInclude Include="..\engine\report\sc_highchart.hpp" />
 		<ClInclude Include="..\engine\sc_enums.hpp" />
 		<ClInclude Include="..\engine\sim\benefit.hpp" />
 		<ClInclude Include="..\engine\sim\cooldown_waste_data.hpp" />
@@ -305,15 +305,15 @@ To change the list of source files run synchronize.py
 		<ClCompile Include="..\engine\report\charts.cpp" />
 		<ClCompile Include="..\engine\report\color.cpp" />
 		<ClCompile Include="..\engine\report\decorators.cpp" />
+		<ClCompile Include="..\engine\report\gear_weights.cpp" />
+		<ClCompile Include="..\engine\report\highchart.cpp" />
 		<ClCompile Include="..\engine\report\json\report_configuration.cpp" />
 		<ClCompile Include="..\engine\report\json\report_json.cpp" />
 		<ClCompile Include="..\engine\report\report_helper.cpp" />
+		<ClCompile Include="..\engine\report\report_html_player.cpp" />
+		<ClCompile Include="..\engine\report\report_html_sim.cpp" />
+		<ClCompile Include="..\engine\report\report_text.cpp" />
 		<ClCompile Include="..\engine\report\reports.cpp" />
-		<ClCompile Include="..\engine\report\sc_gear_weights.cpp" />
-		<ClCompile Include="..\engine\report\sc_highchart.cpp" />
-		<ClCompile Include="..\engine\report\sc_report_html_player.cpp" />
-		<ClCompile Include="..\engine\report\sc_report_html_sim.cpp" />
-		<ClCompile Include="..\engine\report\sc_report_text.cpp" />
 		<ClCompile Include="..\engine\sim\cooldown_waste_data.cpp" />
 		<ClCompile Include="..\engine\sim\event_manager.cpp" />
 		<ClCompile Include="..\engine\sim\proc.cpp" />
