+++ source_files/cmake_engine.txt
@@ -126,11 +126,11 @@ report/charts.hpp
 report/color.hpp
 report/decorators.hpp
 report/gear_weights.hpp
+report/highchart.hpp
 report/json/report_configuration.hpp
 report/report_helper.hpp
 report/report_timer.hpp
 report/reports.hpp
-report/sc_highchart.hpp
 sc_enums.hpp
 sim/benefit.hpp
 sim/cooldown_waste_data.hpp
@@ -299,15 +299,15 @@ player/unique_gear_shadowlands.cpp
 report/charts.cpp
 report/color.cpp
 report/decorators.cpp
+report/gear_weights.cpp
+report/highchart.cpp
 report/json/report_configuration.cpp
 report/json/report_json.cpp
 report/report_helper.cpp
+report/report_html_player.cpp
+report/report_html_sim.cpp
+report/report_text.cpp
 report/reports.cpp
-report/sc_gear_weights.cpp
-report/sc_highchart.cpp
-report/sc_report_html_player.cpp
-report/sc_report_html_sim.cpp
-report/sc_report_text.cpp
 sim/cooldown_waste_data.cpp
 sim/event_manager.cpp
 sim/proc.cpp
