+++ source_files/engine_make
@@ -123,15 +123,15 @@ SRC += \
     report$(PATHSEP)charts.cpp \
     report$(PATHSEP)color.cpp \
     report$(PATHSEP)decorators.cpp \
+    report$(PATHSEP)gear_weights.cpp \
+    report$(PATHSEP)highchart.cpp \
     report$(PATHSEP)json$(PATHSEP)report_configuration.cpp \
     report$(PATHSEP)json$(PATHSEP)report_json.cpp \
     report$(PATHSEP)report_helper.cpp \
+    report$(PATHSEP)report_html_player.cpp \
+    report$(PATHSEP)report_html_sim.cpp \
+    report$(PATHSEP)report_text.cpp \
     report$(PATHSEP)reports.cpp \
-    report$(PATHSEP)sc_gear_weights.cpp \
-    report$(PATHSEP)sc_highchart.cpp \
-    report$(PATHSEP)sc_report_html_player.cpp \
-    report$(PATHSEP)sc_report_html_sim.cpp \
-    report$(PATHSEP)sc_report_text.cpp \
     sim$(PATHSEP)cooldown_waste_data.cpp \
     sim$(PATHSEP)event_manager.cpp \
     sim$(PATHSEP)proc.cpp \
